Build 3204 - Version 3.21.0

+++ CumulusMX/Cumulus.cs
@@ -227,6 +227,7 @@ namespace CumulusMX
 		private static readonly TraceListener FtpTraceListener = new TextWriterTraceListener("ftplog.txt", "ftplog");
 
 		public volatile int WebUpdating;
+		public volatile bool SqlCatchingUp;
 
 		public double WindRoseAngle { get; set; }
 
@@ -490,8 +491,8 @@ namespace CumulusMX
 		private List<string> OWMList = new List<string>();
 
 		// Use thread safe queues for the MySQL command lists
-		private ConcurrentQueue<string> MySqlList = new ConcurrentQueue<string>();
-		public ConcurrentQueue<string> MySqlFailedList = new ConcurrentQueue<string>();
+		private ConcurrentQueue<SqlCache> MySqlList = new ConcurrentQueue<SqlCache>();
+		public ConcurrentQueue<SqlCache> MySqlFailedList = new ConcurrentQueue<SqlCache>();
 
 		// Calibration settings
 		/// <summary>
@@ -7142,7 +7143,7 @@ namespace CumulusMX
 				else
 				{
 					// save the string for later
-					MySqlList.Enqueue(queryString);
+					MySqlList.Enqueue(new SqlCache() { statement = queryString });
 				}
 			}
 		}
@@ -8160,7 +8161,7 @@ namespace CumulusMX
 
 		public void DoLocalCopy()
 		{
-			var remotePath = "";
+			var remotePath = FtpOptions.LocalCopyFolder;
 
 			try
 			{
@@ -8171,7 +8172,7 @@ namespace CumulusMX
 
 				if (FtpOptions.LocalCopyFolder.Length > 0)
 				{
-					remotePath = (FtpOptions.LocalCopyFolder[-1] == DirectorySeparator || FtpOptions.LocalCopyFolder[-1] == folderSep2) ? FtpOptions.LocalCopyFolder : FtpOptions.LocalCopyFolder + DirectorySeparator;
+					remotePath = (FtpOptions.LocalCopyFolder.EndsWith(DirectorySeparator.ToString()) || FtpOptions.LocalCopyFolder.EndsWith(folderSep2.ToString())) ? FtpOptions.LocalCopyFolder : FtpOptions.LocalCopyFolder + DirectorySeparator;
 				}
 				else
 				{
@@ -8812,7 +8813,7 @@ namespace CumulusMX
 		// Return False if the connection is disposed, null, or not connected
 		private bool UploadFile(FtpClient conn, string localfile, string remotefile, int cycle = -1)
 		{
-			string remotefilename = FTPRename ? remotefile + "tmp" : remotefile;
+			string remotefiletmp = FTPRename ? remotefile + "tmp" : remotefile;
 			string cycleStr = cycle >= 0 ? cycle.ToString() : "Int";
 
 			if (FtpOptions.Logging) FtpTrace.WriteLine("");
@@ -8824,13 +8825,36 @@ namespace CumulusMX
 					return true;
 				}
 
+				if (FTPRename)
+				{
+					// delete the existing tmp file
+					try
+					{
+						if (conn.FileExists(remotefiletmp))
+						{
+							conn.DeleteFile(remotefiletmp);
+						}
+					}
+					catch
+					{
+						// continue on error
+					}
+				}
+				
 				if (DeleteBeforeUpload)
 				{
 					// delete the existing file
 					try
 					{
 						LogFtpDebugMessage($"FTP[{cycleStr}]: Deleting {remotefile}");
-						conn.DeleteFile(remotefile);
+						if (conn.FileExists(remotefile))
+						{
+							conn.DeleteFile(remotefile);
+						}
+						else
+						{
+							LogFtpDebugMessage($"FTP[{cycleStr}]: Cannot delete remote file {remotefile} as it does not exist");
+						}
 					}
 					catch (Exception ex)
 					{
@@ -8839,13 +8863,13 @@ namespace CumulusMX
 						{
 							LogFtpMessage($"FTP[{cycleStr}]: Inner Exception: {ex.GetBaseException().Message}");
 						}
-						return conn.IsConnected;
+						// continue on error
 					}
 				}
 
-				LogFtpDebugMessage($"FTP[{cycleStr}]: Uploading {localfile} to {remotefilename}");
+				LogFtpDebugMessage($"FTP[{cycleStr}]: Uploading {localfile} to {remotefiletmp}");
 
-				var status = conn.UploadFile(localfile, remotefilename);
+				var status = conn.UploadFile(localfile, remotefiletmp);
 
 				if (status.IsFailure())
 				{
@@ -8854,16 +8878,16 @@ namespace CumulusMX
 				else if (FTPRename)
 				{
 					// rename the file
-					LogFtpDebugMessage($"FTP[{cycleStr}]: Renaming {remotefilename} to {remotefile}");
+					LogFtpDebugMessage($"FTP[{cycleStr}]: Renaming {remotefiletmp} to {remotefile}");
 
 					try
 					{
-						conn.Rename(remotefilename, remotefile);
-						LogFtpDebugMessage($"FTP[{cycleStr}]: Renamed {remotefilename}");
+						conn.Rename(remotefiletmp, remotefile);
+						LogFtpDebugMessage($"FTP[{cycleStr}]: Renamed {remotefiletmp}");
 					}
 					catch (Exception ex)
 					{
-						LogFtpMessage($"FTP[{cycleStr}]: Error renaming {remotefilename} to {remotefile} : {ex.Message}");
+						LogFtpMessage($"FTP[{cycleStr}]: Error renaming {remotefiletmp} to {remotefile} : {ex.Message}");
 						if (ex.InnerException != null)
 						{
 							LogFtpMessage($"FTP[{cycleStr}]: Inner Exception: {ex.GetBaseException().Message}");
@@ -9297,7 +9321,7 @@ namespace CumulusMX
 			else
 			{
 				// not live, buffer the command for later
-				MySqlList.Enqueue(cmds[0]);
+				MySqlList.Enqueue(new SqlCache() { statement = cmds[0] });
 			}
 		}
 
@@ -9639,6 +9663,9 @@ namespace CumulusMX
 			{
 				if (!MySqlFailedList.IsEmpty)
 				{
+					// flag we are processing the queue so the next task doesn't try as well
+					SqlCatchingUp = true;
+
 					LogMessage($"{callingFunction}: Failed MySQL updates are present");
 					if (MySqlCheckConnection())
 					{
@@ -9654,14 +9681,21 @@ namespace CumulusMX
 						LogMessage($"{callingFunction}: Connection to MySQL server has failed, adding this update to the failed list");
 						if (callingFunction.StartsWith("Realtime["))
 						{
+							var tmp = new SqlCache() { statement = cmds[0] };
+							_ = station.RecentDataDb.Insert(tmp);
+
 							// don't bother buffering the realtime deletes - if present
-							MySqlFailedList.Enqueue(cmds[0]);
+							MySqlFailedList.Enqueue(tmp);
 						}
 						else
 						{
 							for (var i = 0; i < cmds.Count; i++)
 							{
-								MySqlFailedList.Enqueue(cmds[i]);
+								var tmp = new SqlCache() { statement = cmds[i] };
+
+								_ = station.RecentDataDb.Insert(tmp);
+
+								MySqlFailedList.Enqueue(tmp);
 							}
 						}
 					}
@@ -9669,6 +9703,7 @@ namespace CumulusMX
 					{
 						connectionOK = false;
 					}
+					SqlCatchingUp = false;
 				}
 
 				// now do what we came here to do
@@ -9680,6 +9715,7 @@ namespace CumulusMX
 			catch (Exception ex)
 			{
 				LogMessage($"{callingFunction}: Error - " + ex.Message);
+				SqlCatchingUp = false;
 			}
 		}
 
@@ -10158,23 +10194,27 @@ namespace CumulusMX
 
 		public async Task MySqlCommandAsync(string Cmd, string CallingFunction)
 		{
-			var Cmds = new ConcurrentQueue<string>();
-			Cmds.Enqueue(Cmd);
+			var Cmds = new ConcurrentQueue<SqlCache>();
+			Cmds.Enqueue(new SqlCache() { statement = Cmd });
 			await MySqlCommandAsync(Cmds, CallingFunction, false);
 		}
 
 		public async Task MySqlCommandAsync(List<string> Cmds, string CallingFunction)
 		{
-			var tempQ = new ConcurrentQueue<string>(Cmds);
+			var tempQ = new ConcurrentQueue<SqlCache>();
+			foreach (var cmd in Cmds)
+			{
+				tempQ.Enqueue(new SqlCache() { statement = cmd });
+			}
 			await MySqlCommandAsync(tempQ, CallingFunction, false);
 		}
 
-		public async Task MySqlCommandAsync(ConcurrentQueue<string> Cmds, string CallingFunction, bool UseFailedList)
+		public async Task MySqlCommandAsync(ConcurrentQueue<SqlCache> Cmds, string CallingFunction, bool UseFailedList)
 		{
 			await Task.Run(() =>
 			{
-				ConcurrentQueue<string> queue = UseFailedList ? ref MySqlFailedList : ref Cmds;
-				var cmdStr = "";
+				var queue = UseFailedList ? ref MySqlFailedList : ref Cmds;
+				SqlCache cachedCmd = null;
 
 				try
 				{
@@ -10187,11 +10227,11 @@ namespace CumulusMX
 							do
 							{
 								// Do not remove the item from the stack until we know the command worked
-								if (queue.TryPeek(out cmdStr))
+								if (queue.TryPeek(out cachedCmd))
 								{
-									using (MySqlCommand cmd = new MySqlCommand(cmdStr, mySqlConn))
+									using (MySqlCommand cmd = new MySqlCommand(cachedCmd.statement, mySqlConn))
 									{
-										LogDebugMessage($"{CallingFunction}: MySQL executing - {cmdStr}");
+										LogDebugMessage($"{CallingFunction}: MySQL executing - {cachedCmd.statement}");
 
 										if (transaction != null)
 										{
@@ -10201,8 +10241,14 @@ namespace CumulusMX
 										int aff = cmd.ExecuteNonQuery();
 										LogDebugMessage($"{CallingFunction}: MySQL {aff} rows were affected.");
 
-										// Success, pop the value from the queue
-										queue.TryDequeue(out cmdStr);
+										// Success, if using the failed list, delete from the database
+										if (UseFailedList)
+										{
+											station.RecentDataDb.Delete<SqlCache>(cachedCmd.key);
+										}
+
+										// and pop the value from the queue
+										queue.TryDequeue(out cachedCmd);
 									}
 								}
 							} while (!queue.IsEmpty);
@@ -10228,7 +10274,7 @@ namespace CumulusMX
 					// if debug logging is disable, then log the failing statement anyway
 					if (!DebuggingEnabled)
 					{
-						LogMessage($"{CallingFunction}: SQL = {cmdStr}");
+						LogMessage($"{CallingFunction}: SQL = {cachedCmd.statement}");
 					}
 
 
@@ -10249,15 +10295,15 @@ namespace CumulusMX
 
 		public void MySqlCommandSync(string Cmd, string CallingFunction)
 		{
-			var Cmds = new ConcurrentQueue<string>();
-			Cmds.Enqueue(Cmd);
+			var Cmds = new ConcurrentQueue<SqlCache>();
+			Cmds.Enqueue(new SqlCache() { statement = Cmd });
 			MySqlCommandSync(Cmds, CallingFunction, false);
 		}
 
-		public void MySqlCommandSync(ConcurrentQueue<string> Cmds, string CallingFunction, bool UseFailedList)
+		public void MySqlCommandSync(ConcurrentQueue<SqlCache> Cmds, string CallingFunction, bool UseFailedList)
 		{
-			ConcurrentQueue<string> queue = UseFailedList ? ref MySqlFailedList : ref Cmds;
-			var cmdStr = "";
+			var queue = UseFailedList ? ref MySqlFailedList : ref Cmds;
+			SqlCache cachedCmd = null;
 
 			try
 			{
@@ -10270,12 +10316,12 @@ namespace CumulusMX
 						do
 						{
 							// Do not remove the item from the stack until we know the command worked
-							if (queue.TryPeek(out cmdStr))
+							if (queue.TryPeek(out cachedCmd))
 							{
 
-								using (MySqlCommand cmd = new MySqlCommand(cmdStr, mySqlConn))
+								using (MySqlCommand cmd = new MySqlCommand(cachedCmd.statement, mySqlConn))
 								{
-									LogDebugMessage($"{CallingFunction}: MySQL executing - {cmdStr}");
+									LogDebugMessage($"{CallingFunction}: MySQL executing - {cachedCmd.statement}");
 
 									if (transaction != null)
 									{
@@ -10285,8 +10331,13 @@ namespace CumulusMX
 									int aff = cmd.ExecuteNonQuery();
 									LogDebugMessage($"{CallingFunction}: MySQL {aff} rows were affected.");
 
-									// Success, pop the value from the stack
-									queue.TryDequeue(out cmdStr);
+									// Success, if using the failed list, delete from the databasec
+									if (UseFailedList)
+									{
+										station.RecentDataDb.Delete<SqlCache>(cachedCmd.key);
+									}
+									// and pop the value from the queue
+									queue.TryDequeue(out cachedCmd);
 								}
 
 								MySqlUploadAlarm.Triggered = false;
@@ -10312,7 +10363,7 @@ namespace CumulusMX
 				// if debug logging is disable, then log the failing statement anyway
 				if (!DebuggingEnabled)
 				{
-					LogMessage($"{CallingFunction}: SQL = {cmdStr}");
+					LogMessage($"{CallingFunction}: SQL = {cachedCmd}");
 				}
 
 				MySqlUploadAlarm.LastError = ex.Message;
@@ -10333,7 +10384,7 @@ namespace CumulusMX
 			}
 		}
 
-		internal void MySqlCommandErrorHandler(string CallingFunction, int ErrorCode, ConcurrentQueue<string> Cmds)
+		internal void MySqlCommandErrorHandler(string CallingFunction, int ErrorCode, ConcurrentQueue<SqlCache> Cmds)
 		{
 			var ignore = ErrorCode == (int) MySqlErrorCode.ParseError ||
 						 ErrorCode == (int) MySqlErrorCode.EmptyQuery ||
@@ -10356,11 +10407,20 @@ namespace CumulusMX
 			{
 				while (!Cmds.IsEmpty)
 				{
-					Cmds.TryDequeue(out var cmd);
-					if (!cmd.StartsWith("DELETE IGNORE FROM"))
+					try
 					{
-						LogDebugMessage($"{CallingFunction}: Buffering command to failed list");
-						MySqlFailedList.Enqueue(cmd);
+						Cmds.TryDequeue(out var cmd);
+						if (!cmd.statement.StartsWith("DELETE IGNORE FROM"))
+						{
+							LogDebugMessage($"{CallingFunction}: Buffering command to failed list");
+
+							_ = station.RecentDataDb.Insert(cmd);
+							MySqlFailedList.Enqueue(cmd);
+						}
+					}
+					catch (Exception ex)
+					{
+						LogMessage($"{CallingFunction}: Error buffering command - " + ex.Message);
 					}
 				}
 			}
